Create repository and connect-four.py
Added README.md and file structure, and connect-four.py which contains the Board data structure, function for drawing the board to the game screen, game constants, and the main game loop. Currently tracks location of the mouse and hovers a piece above the board, but the game is not yet playable.